requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.3.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
Django==4.2
djangorestframework==3.14.0
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
pycodestyle==2.10.0
pytest==7.4.0
pytest-testdox==3.1.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==2.0.4
